[JBIDE-26380] lookup connection via resource when tailing pod log
(was: lookup via pod)
